Move dir and time funcs around
