+++ c-scape/src/fs/dir/dirfd.rs
@@ -2,7 +2,7 @@ use rustix::fd::{AsFd, AsRawFd};
 
 use libc::{c_int, c_void};
 
-use crate::fs::opendir::MustangDir;
+use super::MustangDir;
 
 #[no_mangle]
 unsafe extern "C" fn dirfd(dir: *mut c_void) -> c_int {
